+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add dungeon super portal block, allowing survival players to select the dungeon they want to do
 - Display Run & Cooldown timers on portal screen
 - Display dead player names in red with "(dead)" suffix on portal screen if allowRespawn is false
+- Better support of modded bosses (now checks if the boss entity is still alive or not)
 ### Fixed:
 - Fixed Dungeon Compass out of bound crash with fancymenu (Thanks listnt)
 - Hybrid NBT + File Data Management for Portal block entity